requirements.txt
# WhiskerBond — Streamlit Cloud compatible requirements
# CPU-only torch — the cloud has no GPU and the CUDA builds are 2GB+
# These exact versions are tested on Streamlit Cloud (Python 3.11)

--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.1.2+cpu
torchvision==0.16.2+cpu

Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.28.0
streamlit>=1.28.0
